Moved some fucntions around added sorting to dir listing.
