+++ server/template.go
@@ -6,11 +6,12 @@ const deftmpl = `<!DOCTYPE html>
 <head>
   <meta charset="utf-8">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
+	<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.4.1/css/all.css" integrity="sha384-5sAR7xN1Nv6T6+dT2mhtzEpVJvfS3NScPQTrOxhwjIuvcA67KV2R5Jz6kr4abQsz" crossorigin="anonymous">
 </head>
 	<body>
 	 <div class="container">
-	 <table class="centered">
+	 <table class="responsive-table">
 	  <thead>
 	    <tr>
 		   <th>Name</th>
@@ -20,11 +21,15 @@ const deftmpl = `<!DOCTYPE html>
 	  </thead>
 	<tbody>
 	 <tr>
-	    <td><a href="../">Go Back...</a></td>
+	    <td><i class="fas fa-reply"></i><a href="../">  Go Back...</a></td>
 	 </tr>
 	{{range .Items}}
-	  <tr>
-	    <td><a href="{{.Name}}">{{.Name}}</a></td>
+		<tr>
+			{{if .Dir}}
+       <td><i class="fas fa-folder-open"></i><a href="{{.Name}}"> {{.Name}}</a></td>
+			{{else}}
+			 <td><i class="fas fa-file"></i><a href="{{.Name}}"> {{.Name}}</a></td>
+			{{end}}
 	    <td><a>{{.LastMod}}</a></td>
 	    <td><a>{{.Size}}</a></td>
 	  </tr>
